+++ packages/core/platform/tests/platform-specs.js
@@ -104,6 +104,17 @@ describe('platform', () => {
 		});
 	});
 
+	describe('parseUserAgent for Windows Phone', () => {
+		const windowsPhone = 'Mozilla/5.0 (Windows Phone 8.0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4103.84 Mobile Safari/537.36';
+
+		test('should return platformName `windowsPhone`', () => {
+			const expected = {platformName: 'windowsPhone'};
+			const actual = parseUserAgent(windowsPhone);
+
+			expect(actual).toMatchObject(expected);
+		});
+	});
+
 	describe('parseUserAgent for User-Agent Reduction', () => {
 		const testVersion = '113';
 		const uaGenerator = (unifiedPlatform, deviceCompatibility = '', majorVersion = testVersion) => (